+++ src/style/style_layer/raster_style_layer.ts
@@ -1,4 +1,5 @@
 import {StyleLayer} from '../style_layer';
+import {warnOnce} from '../../util/util';
 
 import properties, {type RasterPaintPropsPossiblyEvaluated} from './raster_style_layer_properties.g';
 import {type Transitionable, type Transitioning, type PossiblyEvaluated} from '../properties';
@@ -14,6 +15,10 @@ export class RasterStyleLayer extends StyleLayer {
     paint: PossiblyEvaluated<RasterPaintProps, RasterPaintPropsPossiblyEvaluated>;
 
     constructor(layer: LayerSpecification, globalState: Record<string, any>) {
+        const resampling = layer.paint?.['resampling'];
+        if (resampling && layer.paint?.['raster-resampling']) {
+            warnOnce(`Raster layer "${layer.id}" paint properties "resampling" and "raster-resampling" are both specified, but only "resampling" needs to be specified. Defaulting to "resampling" (${resampling}).`);
+        }
         super(layer, properties, globalState);
     }
 }